Move remaining store files to entities directory.
