+++ src/core/entities/configDocuments.js
